Moved matrix stuff to its own file
